No more soft deletion, Handle already in transaction
I totally forgot we have SyncedEntity to track deletions of objects, hence we don't really need a `isDeleted` property for objects. But a new method `realm.delete(cloudableObject:)` is needed to convey changes to SyncedEntity on deletions.

Notification blocks can be triggered inside a write transaction, so we have to guard it.